Add files via upload
Slide-deck for the first ME-NET stakeholder engagement board meeting, including the first iteration of visual educational content co-produced with Common Knowledge for communicating the impact of ozone on health outcomes to divers stakeholders. Stakeholder feedback is captured in meeting transcripts. Feedback informed consecutive iterations of educational content, in addition to ME-NET phone and web application features and functions, including self-reporting mental health, the design of an interactive climate and health dashboard for visualising the effect of methane on health outcomes via ozone, and educational 'modules'.